requirements.txt
Babel==2.1.1
Jinja2==2.8
MarkupSafe==0.23
Pygments==2.0.2
Sphinx==1.3.1
Twisted==15.4.0
alabaster==0.7.6
construct==2.5.2
coverage==4.0.1
docutils==0.12
pluggy==0.3.1
py==1.4.30
pytest==2.8.2
pytest-cov==2.2.0
pytest-pythonpath==0.7
pytz==2015.7
repoze.sphinx.autointerface==0.7.1
six==1.10.0
snowballstemmer==1.2.0
sphinx-rtd-theme==0.1.9
tox==2.1.1
virtualenv==13.1.2
wsgiref==0.1.2
zope.interface==4.1.3
